added react-native-navigation and wrapped Home.js inside App.js
