+++ Home.js
@@ -1,0 +1,34 @@
+import * as React from 'react';
+import { Platform, StyleSheet, Text, View } from 'react-native';
+
+const instructions = Platform.select({
+  ios: `Press Cmd+R to reload,\nCmd+D or shake for dev menu`,
+  android: `Double tap R on your keyboard to reload,\nShake or press menu button for dev menu`,
+});
+
+export default function Home() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.welcome}>Welcome to Muze!</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#F5FCFF',
+  },
+  welcome: {
+    fontSize: 20,
+    textAlign: 'center',
+    margin: 10,
+  },
+  instructions: {
+    textAlign: 'center',
+    color: '#333333',
+    marginBottom: 5,
+  },
+});
