+++ ios/muze/Supporting/Expo.plist
@@ -2,9 +2,11 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-        <key>EXUpdatesSDKVersion</key>
-        <string>YOUR-APP-SDK-VERSION-HERE</string>
-        <key>EXUpdatesURL</key>
-        <string>YOUR-APP-URL-HERE</string>
+	<key>EXUpdatesReleaseChannel</key>
+	<string>default</string>
+	<key>EXUpdatesSDKVersion</key>
+	<string>37.0.0</string>
+	<key>EXUpdatesURL</key>
+	<string>https://exp.host/@heyabird/muze</string>
 </dict>
 </plist>
